refactor(whispering): route pill gestures through one pill-actions module
The "translate a pill gesture into a recorder operation" rule was duplicated:
the web host branched on capture in handleStop, and the Tauri main window
branched again in its action listener. The two had already drifted (web's
cancel was ungated, relying on the pill only rendering a cancel button in
manual mode; Tauri gated it to manual capture).

Collapse both into one dispatchPillAction, co-located with the failed-pill
open-detail gesture in a new pill-actions module that replaces focus-failure
(a stale name: it never focused, and with the latch it now clears state). Both
pill mounts import two functions from one file instead of five from two, the
Tauri action listener drops to a single call, and cancel is gated to manual in
one place.
