+++ apps/whispering/src/lib/recording-overlay/RecordingPillHost.svelte
@@ -1,12 +1,9 @@
 <script lang="ts">
 	import { tauri } from '#platform/tauri';
 	import {
-		cancelRecording,
-		retryDictation,
-		stopManualRecording,
-		stopVadRecording,
-	} from '$lib/operations/recording';
-	import { openFailedDictationDetail } from '$lib/recording-overlay/focus-failure';
+		dispatchPillAction,
+		openFailedDictationDetail,
+	} from '$lib/recording-overlay/pill-actions';
 	import RecordingPill from '$lib/recording-overlay/RecordingPill.svelte';
 	import { projectLifecycleToStatus } from '$lib/recording-overlay/projection';
 	import { webPillLevel } from '$lib/recording-overlay/web-pill.svelte';
@@ -15,30 +12,10 @@
 	// The web mount of the shared dictation pill. On desktop the pill is a native
 	// overlay window, so this host renders nothing there; on web it places the
 	// same `RecordingPill` as a fixed bottom-center element and drives it straight
-	// from the lifecycle value, calling the recorder operations directly (no IPC).
+	// from the lifecycle value, routing gestures through `pill-actions` (no IPC).
+	// On web the app window is already focused, so the body click only opens the
+	// failed recording's row; there is no window to raise.
 	const status = $derived(projectLifecycleToStatus(dictationLifecycle.current));
-
-	function handleStop() {
-		// Stop acts on whichever capture is live.
-		const { capture } = dictationLifecycle.current;
-		if (capture.kind !== 'recording') return;
-		if (capture.trigger === 'manual') void stopManualRecording();
-		else void stopVadRecording();
-	}
-
-	function handleCancel() {
-		void cancelRecording();
-	}
-
-	function handleRetry() {
-		void retryDictation();
-	}
-
-	function handleFocusMain() {
-		// On web the app window is already the focused surface, so there is no
-		// window to raise; a failure still opens the recording's row.
-		openFailedDictationDetail();
-	}
 </script>
 
 {#if !tauri && status}
@@ -46,10 +23,10 @@
 		<RecordingPill
 			{status}
 			level={webPillLevel.level}
-			onStop={handleStop}
-			onCancel={handleCancel}
-			onRetry={handleRetry}
-			onFocusMain={handleFocusMain}
+			onStop={() => dispatchPillAction('stop')}
+			onCancel={() => dispatchPillAction('cancel')}
+			onRetry={() => dispatchPillAction('retry')}
+			onFocusMain={openFailedDictationDetail}
 		/>
 	</div>
 {/if}
